# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
- url: /v0
paths:
  /nexuses:
    get:
      operationId: get_nexuses
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Nexus'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
  /nexuses/{nexus_id}:
    delete:
      operationId: del_nexus
      parameters:
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
    get:
      operationId: get_nexus
      parameters:
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Nexus'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
...
  /nodes/{id}/nexuses:
    get:
      operationId: get_node_nexuses
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Nexus'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
...
  /nodes/{node_id}/nexuses/{nexus_id}:
    delete:
      operationId: del_node_nexus
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
    get:
      operationId: get_node_nexus
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Nexus'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
    put:
      operationId: put_node_nexus
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNexusBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Nexus'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
...
  /nodes/{node_id}/nexuses/{nexus_id}/share:
    delete:
      operationId: del_node_nexus_shares
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
  /nodes/{node_id}/nexuses/{nexus_id}/share/{protocol}:
    put:
      operationId: put_node_nexus_share
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: nexus_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: protocol
        required: true
        schema:
          enum:
          - nvmf
          - iscsi
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Nexuses
...
components:
  schemas:
...
    CreateNexusBody:
      description: Create Nexus Body JSON
      example:
        children:
        - ""
        size: 0
      properties:
        children:
          description: |-
            replica can be iscsi and nvmf remote targets or a local spdk bdev
             (i.e. bdev:///name-of-the-bdev).

             uris to the targets we connect to
          items:
            type: string
          type: array
        size:
          description: size of the device in bytes
          format: int64
          type: integer
      required:
      - children
      - size
      type: object
...
    Nexus:
      description: Nexus information
      example:
        children:
        - rebuildProgress: null
          state: Unknown
          uri: ""
        deviceUri: ""
        node: ""
        rebuilds: 0
        share: "off"
        size: 0
        state: Unknown
        uuid: 514ed1c8-7174-49ac-b9cd-ad44ef670a67
      properties:
        children:
          description: array of children
          items:
            $ref: '#/components/schemas/Nexus_children'
          type: array
        deviceUri:
          description: |-
            URI of the device for the volume (missing if not published).
             Missing property and empty string are treated the same.
          type: string
        node:
          description: id of the mayastor instance
          type: string
        rebuilds:
          description: total number of rebuild tasks
          format: int32
          type: integer
        share:
          description: protocol used for exposing the nexus
          enum:
          - "false"
          - nvmf
          - iscsi
          - nbd
          type: string
        size:
          description: size of the volume in bytes
          format: int64
          type: integer
        state:
          description: current state of the nexus
          enum:
          - Unknown
          - Online
          - Degraded
          - Faulted
          type: string
        uuid:
          description: uuid of the nexus
          format: uuid
          type: string
      required:
      - children
      - deviceUri
      - node
      - rebuilds
      - share
      - size
      - state
      - uuid
      type: object
...
    RestJsonError:
      description: Rest Json Error format
      example:
        details: ""
        error: NotFound
      properties:
        details:
          description: detailed error information
          type: string
        error:
          description: error kind
          enum:
          - Timeout
          - Deserialize
          - Internal
          - InvalidArgument
          - DeadlineExceeded
          - NotFound
          - AlreadyExists
          - PermissionDenied
          - ResourceExhausted
          - FailedPrecondition
          - NotShared
          - NotPublished
          - AlreadyPublished
          - AlreadyShared
          - Aborted
          - OutOfRange
          - Unimplemented
          - Unavailable
          - Unauthenticated
          - Unauthorized
          - Conflict
          - FailedPersist
          - Deleting
          type: string
      required:
      - details
      - error
      type: object
...
    Nexus_children:
      description: Child information
      example:
        rebuildProgress: null
        state: Unknown
        uri: ""
      properties:
        rebuildProgress:
          description: current rebuild progress (%)
          format: int32
          type: integer
        state:
          description: state of the child
          enum:
          - Unknown
          - Online
          - Degraded
          - Faulted
          type: string
        uri:
          description: uri of the child device
          type: string
      required:
      - state
      - uri
      type: object
...
  securitySchemes:
    JWT:
      bearerFormat: JWT
      scheme: bearer
      type: http